+++ demisto_sdk/__main__.py
@@ -20,8 +20,8 @@ from demisto_sdk.commands.common.tools import (filter_files_by_type,
                                                get_pack_name, print_error,
                                                print_warning)
 from demisto_sdk.commands.common.update_id_set import merge_id_sets_from_files
-from demisto_sdk.commands.create_artifacts.content_artifacts_creator import (
-    ArtifactsManager, create_content_artifacts)
+from demisto_sdk.commands.create_artifacts.content_artifacts_creator import \
+    ArtifactsManager
 from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
 from demisto_sdk.commands.download.downloader import Downloader
 from demisto_sdk.commands.find_dependencies.find_dependencies import \
@@ -244,7 +244,7 @@ def unify(**kwargs):
 )
 @click.option(
     '-a', '--validate-all', is_flag=True, show_default=True, default=False,
-    help='Whether to run all validation on all files or not'
+    help='Whether to run all validation on all files or not.'
 )
 @click.option(
     '-i', '--input', type=click.Path(exists=True), help='The path of the content pack/file to validate specifically.'
@@ -319,7 +319,8 @@ def validate(config, **kwargs):
                '1. content_new - Contains all content objects of type json,yaml (from_version < 6.0.0)'
                '2. content_packs - Contains all packs from Packs - Ignoring internal files (to_version >= 6.0.0).'
                '3. content_test - Contains all test scripts/playbooks (from_version < 6.0.0)'
-               '4. content_all - Contains all from content_new and content_test.')
+               '4. content_all - Contains all from content_new and content_test.'
+               '5. uploadable_packs - Contains zipped packs that are ready to be uploaded to Cortex XSOAR machine.')
 @click.help_option('-h', '--help')
 @click.option('-a', '--artifacts_path', help='Destination directory to create the artifacts.',
               type=click.Path(file_okay=False, resolve_path=True), required=True)
@@ -328,11 +329,23 @@ def validate(config, **kwargs):
 @click.option('-v', '--content_version', help='The content version in CommonServerPython.', default='0.0.0')
 @click.option('-s', '--suffix', help='Suffix to add all yaml/json/yml files in the created artifacts.')
 @click.option('--cpus',
-              help='Number of cpus/vcpus availble - only required when os not reflect number of cpus (CircleCI'
-                   'allways show 32, but medium has 3.', hidden=True, default=os.cpu_count())
-def create_arifacts(**kwargs) -> int:
+              help='Number of cpus/vcpus available - only required when os not reflect number of cpus (CircleCI'
+                   'always show 32, but medium has 3.', hidden=True, default=os.cpu_count())
+@click.option('-idp', '--id-set-path', help='The full path of id_set.json', hidden=True,
+              type=click.Path(exists=True, resolve_path=True))
+@click.option('-p', '--pack-names',
+              help=("Packs to create artifacts for. Optional values are: `all` or "
+                    "csv list of packs. "
+                    "Default is set to `all`"),
+              default="all", hidden=True)
+@click.option('-sk', '--signature-key', help='Base64 encoded signature key used for signing packs.', hidden=True)
+@click.option('-sd', '--sign-directory', help='Path to the signDirectory executable file.',
+              type=click.Path(exists=True, resolve_path=True), hidden=True)
+@click.option('-rt', '--remove-test-playbooks', is_flag=True,
+              help='Should remove test playbooks from content packs or not.', default=True, hidden=True)
+def create_artifacts(**kwargs) -> int:
     artifacts_conf = ArtifactsManager(**kwargs)
-    return create_content_artifacts(artifacts_conf)
+    return artifacts_conf.create_content_artifacts()
 
 
 # ====================== secrets ====================== #